Parameter WithForeign.Map2

include NODE with type 'a key = 'a key

Types

type 'a key = 'a key

The type of keys.

type ('key, 'map) value

The type of value, which depends on the type of the key and the type of the map.

type 'map t

The type of the map, which is parameterized by a type.

Constructors: build values

val empty : 'map t

The empty map

val leaf : 'key key -> ('key, 'map) value -> 'map t

A singleton leaf, similar to BASE_MAP.singleton

val branch : prefix:int -> branching_bit:int -> tree0:'map t -> tree1:'map t -> 'map t

A branch node. This shouldn't be called externally unless you know what you're doing! Doing so could easily break the data structure's invariants.

When called, it assumes that:

Neither tree0 nor tree1 should be empty.

branching_bit should have a single bit set

prefix should be normalized (bits below branching_bit set to zero)

All elements of tree0 should have their to_int start by prefix followed by 0 at position branching_bit).

All elements of tree1 should have their to_int start by prefix followed by 0 at position branching_bit).

Destructors: access the value

type 'map view = private

| Empty : 'map view
(*
Can happen only at the toplevel: there is no empty interior node.
*)

| Branch : {

prefix : int;

branching_bit : int;

tree0 : 'map t;

tree1 : 'map t;

} -> 'map view
(*
Same constraints as branch:

branching_bit contains only one bit set; the corresponding mask is (branching_bit - 1).

prefix is normalized: the bits below the branching_bit are set to zero (i.e. prefix & (branching_bit - 1) = 0).

All elements of tree0 should have their to_int start by prefix followed by 0 at position branching_bit).

All elements of tree1 should have their to_int start by prefix followed by 0 at position branching_bit).

*)

| Leaf : {

key : 'key key;

value : ('key, 'map) value;

} -> 'map view
(*
A key -> value mapping.
*)

This makes the map nodes accessible to the pattern matching algorithm; this corresponds 1:1 to the SimpleNode implementation. This just needs to be copy-and-pasted for every node type.

val is_empty : 'map t -> bool

Check if the map is empty. Should be constant time.

val view : 'a t -> 'a view

Convert the map to a view. Should be constant time.

val find : 'key key -> 'map t -> ('key, 'map) value

find key map returns the value associated with key in map if present.

raises Not_found
if key is absent from map

val find_opt : 'key key -> 'map t -> ('key, 'map) value option

Same as find, but returns None for Not_found