Module type PatriciaTree.HETEROGENEOUS_VALUE

The module type of values, which can be heterogeneous. This can be used to specify how the type of the value depends on that of the key. If the value doesn't depend on the key type, you can use the provided default implementations HomogeneousValue and WrappedHomogeneousValue.

type ('key, 'map) t

The type of values. A 'map map maps 'key key to ('key, 'map) value. Can be mutable if desired, unless it is being used in Hash-consed maps and sets.