Parameter MakeSet.Key

type t

The type of keys

val to_int : t -> int

A unique identifier for values of the type. Usually, we use a fresh counter that is increased to give a unique id to each object. Correctness of the operations requires that different values in a tree correspond to different integers.

Must be injective, return only positive values, and ideally fast